+++ scripts/install-external-deps.js
@@ -31,13 +31,20 @@ const downloadUrlToDirectory = (url, dir) =>
         .on('finish', resolve)
     }));
 
-const unzip = (source, target) => new Promise((resolve, reject) => {
-  extract(source, { dir: target }, (err) => {
-    if (err) {
-      reject(err);
-    }
-    resolve(target);
-  });
+const unzip = async (source, target) => extract(source, { dir: target });
+const move = async (src, dest) => fs.move(src, dest, { overwrite: true });
+const waitForFile = async(filePath) => new Promise((resolve, reject) => {
+  let responded = false;
+  const done = (error) => {
+    if (responded) return;
+    responded = true;
+    clearInterval(interval);
+    clearTimeout(timeout);
+    return error ? reject(error) : resolve();
+  };
+
+  const interval = setInterval(() => (fs.existsSync(filePath) && done()), 100);
+  const timeout = setTimeout(() => done(`Timeout waiting for file ${filePath}`), 5000);
 });
 
 const downloadChromium = () => {
@@ -71,9 +78,12 @@ const downloadChromedriver = () => {
   const chromedriverFinalPath = path.join(__dirname, '..', 'node_modules', 'chromedriver', 'lib', 'chromedriver', 'chromedriver');
 
   return downloadUrlToDirectory(chromedriverUrl, chromedriverTmpZip)
+    .then(() => waitForFile(chromedriverTmpZip))
     .then(() => unzip(chromedriverTmpZip, browserlessTmpDir))
-    .then(() => fs.move(chromedriverUnzippedPath, chromedriverFinalPath, { overwrite: true }))
-    .then(() => fs.chmodSync(chromedriverFinalPath, '755'));
+    .then(() => waitForFile(chromedriverUnzippedPath))
+    .then(() => move(chromedriverUnzippedPath, chromedriverFinalPath))
+    .then(() => fs.chmod(chromedriverFinalPath, '755'))
+    .then(() => waitForFile(chromedriverFinalPath));
 };
 
 const downloadDevTools = () => {
@@ -83,11 +93,14 @@ const downloadDevTools = () => {
   const devtoolsFinalPath = path.join(__dirname, '..', 'debugger', 'devtools');
 
   return downloadUrlToDirectory(devtoolsUrl, devtoolsTmpZip)
+    .then(() => waitForFile(devtoolsTmpZip))
     .then(() => unzip(devtoolsTmpZip, browserlessTmpDir))
-    .then(() => fs.move(devtoolsUnzippedPath, devtoolsFinalPath, { overwrite: true }))
+    .then(() => waitForFile(devtoolsUnzippedPath))
+    .then(() => move(devtoolsUnzippedPath, devtoolsFinalPath))
+    .then(() => waitForFile(devtoolsFinalPath));
 };
 
-(async () => {
+(() => new Promise(async (resolve, reject) => {
   try {
     await fs.mkdir(browserlessTmpDir);
     await Promise.all([
@@ -95,10 +108,15 @@ const downloadDevTools = () => {
       downloadChromedriver(),
       downloadDevTools(),
     ]);
-    console.log('Done unpacking chromedriver and devtools assets');
   } catch(err) {
-    console.error(`Error unpacking chromedriver and devtools assets:\n${err.message}\n${err.stack}`);
+    console.error(`Error unpacking assets:\n${err.message}\n${err.stack}`);
+    reject(err);
+    process.exit(1);
   } finally {
-    rimraf(browserlessTmpDir, () => {});
+    rimraf(browserlessTmpDir, (err) => {
+      console.log('Done unpacking chromedriver and devtools assets');
+      if (err) console.warn(`Error removing temporary directory ${browserlessTmpDir}`);
+      resolve();
+    });
   }
-})();
+}))();