+++ src/apis/screencast.ts
@@ -1,47 +1,130 @@
-import * as fs from 'fs';
-import { noop } from 'lodash';
-import {
-  IAfter,
-  IBefore,
-  IPreferences,
-} from '../types';
+import * as path from 'path';
+import { after as downloadAfter } from './download';
+import { id, mkdir } from '../utils';
+import { WORKSPACE_DIR } from '../config';
+import { IBefore } from '../types';
 
-const rimraf = require('rimraf');
+export const before = async ({ page, code, debug, browser }: IBefore) => {
+  // @ts-ignore reaching into private methods
+  const client = page._client;
+  const renderer = await browser.newPage();
+  const downloadPath = path.join(WORKSPACE_DIR, `.browserless.download.${id()}`);
+  await mkdir(downloadPath);
+  const downloadName = id() + '.webm';
+  let screencastAPI: any;
 
-export const before = async ({ page, code, debug }: IBefore) => {
-  let prefs: IPreferences;
+  // @ts-ignore
+  await renderer._client.send('Page.setDownloadBehavior', {
+    behavior: 'allow',
+    downloadPath,
+  });
 
-  const startScreencast = () => page.evaluate(() => window.postMessage({ type: 'REC_START' }, '*'));
-  const stopScreencast = () => page.evaluate(() => window.postMessage({ type: 'REC_STOP' }, '*'));
-  const setPreferences = (preferences: IPreferences) => prefs = preferences;
+  // Setup page handlers
+  const setup = async () => await renderer.evaluateHandle((downloadName) => {
+    const screencastAPI = class {
+      private canvas: HTMLCanvasElement;
+      private ctx: CanvasRenderingContext2D;
+      private downloadAnchor: HTMLAnchorElement;
+      private recordingFinish: Promise<void>;
+      private recorder: any;
+      private chunks: any[];
 
-  const setupScreencast = () => page.evaluate(
-    (viewport) => {
-      const { height, width } = JSON.parse(viewport);
-      return window.postMessage({
-        height,
-        type: 'REC_CLIENT_SETUP',
-        width,
-      }, '*');
-    },
-    JSON.stringify(page.viewport()),
-  );
+      constructor() {
+        this.canvas = document.createElement('canvas');
+        this.downloadAnchor = document.createElement('a');
+
+        document.body.appendChild(this.canvas);
+        document.body.appendChild(this.downloadAnchor);
+
+        this.ctx = this.canvas.getContext('2d') as CanvasRenderingContext2D;
+        this.downloadAnchor.href = '#';
+        this.downloadAnchor.textContent = 'Download video';
+        this.downloadAnchor.id = 'download';
+        this.chunks = [];
+      }
+
+      private async beginRecording(stream: any): Promise<any[]> {
+        return new Promise((resolve, reject) => {
+          // @ts-ignore No MediaRecorder
+          this.recorder = new MediaRecorder(stream, { mimeType: 'video/webm' });
+          this.recorder.ondataavailable = (e: any) => this.chunks.push(e.data);
+          this.recorder.onerror = reject;
+          this.recorder.onstop = resolve;
+          this.recorder.start();
+        });
+      }
+
+      private async download() {
+        await this.recordingFinish;
+        const blob = new Blob(this.chunks, { type: 'video/webm' });
+
+        this.downloadAnchor.onclick = () => {
+          this.downloadAnchor.href = URL.createObjectURL(blob);
+          this.downloadAnchor.download = downloadName;
+        };
+
+        this.downloadAnchor.click();
+      }
+
+      async start({ width, height }: { width: number; height: number }) {
+        this.canvas.width = width;
+        this.canvas.height = height;
+        // @ts-ignore No captureStream API
+        this.recordingFinish = this.beginRecording(this.canvas.captureStream());
+      }
+
+      async draw(pngData: Buffer) {
+        const data = await fetch(`data:image/png;base64,${pngData}`)
+          .then(res => res.blob())
+          .then(blob => createImageBitmap(blob));
+
+        this.ctx.clearRect(0, 0, this.canvas.width, this.canvas.height);
+        this.ctx.drawImage(data, 0, 0);
+
+        return this;
+      }
+
+      stop() {
+        this.recorder.stop();
+        this.download();
+        return this;
+      }
+    };
+
+    return new screencastAPI();
+  }, downloadName);
+
+  const startScreencast = async () => {
+    const viewport = page.viewport();
+    screencastAPI = await setup();
+    await page.bringToFront();
+
+    await renderer.evaluateHandle(
+      (screencastAPI, width, height) => screencastAPI.start({ width, height }),
+      screencastAPI,
+      viewport.width, viewport.height
+    );
+
+    await client.send('Page.startScreencast', {
+      format: 'jpeg',
+      maxWidth: viewport.width,
+      maxHeight: viewport.height,
+      everyNthFrame: 1,
+    });
+
+    client.on('Page.screencastFrame', ({ data, sessionId }: { data: string; sessionId: string }) => {
+      renderer.evaluateHandle((screencastAPI, data) => screencastAPI.draw(data), screencastAPI, data);
+      client.send('Page.screencastFrameAck', { sessionId }).catch(() => {});
+    });
+  };
+
+  const stopScreencast = async () => {
+    await client.send('Page.stopScreencast');
+    await renderer.bringToFront();
+    await renderer.evaluateHandle((screencastAPI) => screencastAPI.stop(), screencastAPI);
+  };
 
   page.on('load', async () => {
-    if (prefs) {
-      debug(`Injecting preferences: ${JSON.stringify(prefs, null, '  ')}`);
-      await page.evaluate(
-        (prefs) => window.postMessage({
-          prefs: JSON.parse(prefs),
-          type: 'SET_PREFERENCES',
-        }, '*'),
-        JSON.stringify(prefs),
-      );
-    }
-
-    debug(`Setting up screencast`);
-    await setupScreencast();
-
     if (!code.includes('startScreencast')) {
       debug(`Starting to record`);
       setTimeout(startScreencast, 0);
@@ -49,40 +132,19 @@ export const before = async ({ page, code, debug }: IBefore) => {
   });
 
   return {
-    setPreferences,
+    downloadPath,
     startScreencast,
     stopScreencast,
   };
 };
 
-export const after = async ({ page, res, done, debug, code, stopScreencast }: IAfter) => {
+export const after = async(
+  { code, stopScreencast, downloadPath, debug, res, done }:
+  { code: string; stopScreencast: () => Promise<void>; downloadPath: string, debug: (...args: string[]) => {}, res: any, done: (errBack?: Error | null) => {} },
+) => {
   if (!code.includes('stopScreencast')) {
     await stopScreencast();
   }
 
-  await page.waitForSelector('html.downloadComplete', { timeout: 0 });
-  const filePath = await page.evaluate(() => document.querySelector('html')!.getAttribute('data-filepath'));
-
-  debug(`Screencast download "${filePath}" complete!`);
-
-  if (!filePath || !fs.existsSync(filePath)) {
-    debug(`Couldn't located screencast in the filesystem at "${filePath}"`);
-    throw new Error(`Couldn't locate screencast file "${filePath}"`);
-  }
-
-  if (res.headersSent) {
-    rimraf(filePath, noop);
-    return done();
-  }
-
-  return res.sendFile(filePath, (err: Error) => {
-    const message = err ?
-      `Error streaming file back ${err}` :
-      `File sent successfully`;
-
-    debug(message);
-    rimraf(filePath, noop);
-
-    done(err);
-  });
+  return downloadAfter({ downloadPath, debug, res, done });
 };